gestion_eventos/requirements.txt
Django>=5.2.5
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
